feat: update privacy policy (#98)
Update privacy.md to reflect a privacy policy that governs the use of the ory.sh website. The scope of this new privacy policy is not to include policy for user accounts or other services not yet in existence as part of the ory offering - this will be added later.